stock_agent/requirements.txt
python-dotenv
langgraph
langchain
langchain_openai
langchain_anthropic
langchain_google_genai
langchain-core
pydantic
streamlit
typing
yfinance
finnhub-python
fredapi
langchain-tavily
cachetools
langgraph-cli